+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.10.0 (edge)
 
+* Fix using more than two `rescue` in sequence (#1269)
+
 * Fixed inheritance for Array subclasses.
 
 * Always populate all stub_subscribers with all method stubs, as a side effect of this now `method_missing` on bridged classes now works reliably (#1273)